requirements.txt
groq
numpy
pandas
python-dotenv
streamlit
